Handle invalid user input character
- if a user types in 't', this will display invalid character and reprint the inventory and menu
- add invalid input test

+++ Include/CoffeeMachine.h
@@ -78,7 +78,7 @@ public:
   CoffeeMachine();
   ~CoffeeMachine();
   CoffeeMachine(CoffeeMachine const &other);
-  void makeDrink(int i);
+  void makeDrink(long i);
   bool isStocked(Drink &d);
   double calcPrice(std::string drinkName) const;
   void displayInventory();
